requirements.txt
# Legacy Interview App - Backend Dependencies
agno>=2.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-decouple>=3.8
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.25.0
openai-whisper>=20231117
pydantic>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg[binary]>=3.1.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
redis>=4.0.0

# Audio Processing Dependencies
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.11.0
numpy>=1.24.0
webrtcvad>=2.0.10

# Speaker Diarization - PyAnnote-Audio (replaces complex custom implementation)
pyannote.audio>=3.1.0
torch>=1.9.0  # For GPU support
torchaudio>=0.9.0  # For audio processing

# Audio Evaluation Dependencies
jiwer>=3.0.0  # For Word Error Rate calculations

# Testing Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0
httpx>=0.24.0
factory-boy>=3.3.0
